requirements.txt
asgiref==3.6.0
boto3==1.26.49
botocore==1.29.49
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-writable-nested==0.7.0
drf-yasg==1.21.4
environ==1.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==22.0
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
